flatpak-builder.rst: Make the process standalone
Usually the developer already have some .json file, and if not, is more comfortable anyway to have a file where you can edit etc. than fiddling around with CLI.

In the flatpak-builder tutorial, it refers to a previous section to export the built application to a repo previously created. The previous section uses build-init and in no place it says what is necessary from that section to actually use flatpak-builder, so the developer ends up doing the whole thing instead of just using flatpak-builder with the json file.

In order to fix that, make the flatpak-builder tutorial standalone, ready to be used with no previous set up.